refactor: move components, icons, context, provider, hook under dedicated folders
