+++ frontend/src/features/chat/chat-da/components/ChatHeader/icons/index.ts
@@ -1,2 +1,1 @@
-export { SendIcon } from './SendIcon';
 export { ChatIcon } from './ChatIcon';
