+++ frontend/src/features/chat/chat-da/components/ChatInput/icons/index.ts
@@ -1,0 +1,1 @@
+export { SendIcon } from './SendIcon';
