+++ frontend/src/features/chat/chat-da/components/index.ts
@@ -1,0 +1,3 @@
+export { ChatHeader } from './ChatHeader';
+export { ChatInput } from './ChatInput';
+export { ChatMessage } from './ChatMessage';
