+++ frontend/src/features/chat/chat-da/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useChat } from './useChat';
