+++ frontend/src/features/chat/chat-da/views/index.ts
@@ -1,0 +1,2 @@
+export { FullScreenChat } from './FullScreenChat';
+export { SidebarChat } from './SidebarChat';
